feat: standardize error hierarchy and document recovery patterns for SDK consumers
